# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vokal Director API",
    "description": "Контракт между фронтендом и сервером Vokal Director.\n\n**Что здесь важно знать до чтения адресов.**\n\n* Большая часть адресов отвечает `501`. Это не заготовка на будущее: схема\n  запроса и ответа настоящая и полная, реализации за адресом еще нет. В\n  `details.missing` сказано, чего именно не хватает. Заглушек, возвращающих\n  правдоподобные данные, в API нет — мок в ответе сервера хуже отказа.\n* Ошибка всегда выглядит одинаково: `{\"error\": {\"code\", \"message\",\n  \"requestId\", \"details\"}}`.\n* Имена полей — camelCase. Идентификаторы — непрозрачные строки.\n* Незнакомое поле в теле запроса — ошибка `422`, а не тихо отброшенное\n  значение.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/recipients": {
      "get": {
        "tags": [
          "выдача"
        ],
        "summary": "Получатели материалов",
        "operationId": "list_recipients_api_projects__project_id__recipients_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Сколько записей вернуть",
              "default": 20,
              "title": "Limit"
            },
            "description": "Сколько записей вернуть"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Сколько записей пропустить",
              "default": 0,
              "title": "Offset"
            },
            "description": "Сколько записей пропустить"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareRecipientListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "выдача"
        ],
        "summary": "Добавить получателя",
        "description": "Разным ролям выдается разное: у преподавателя и ученика материалы не совпадают.",
        "operationId": "create_recipient_api_projects__project_id__recipients_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareRecipientCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareRecipientOut"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Данные запроса не прошли проверку схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/recipients/{recipient_id}": {
      "patch": {
        "tags": [
          "выдача"
        ],
        "summary": "Изменить получателя",
        "operationId": "update_recipient_api_projects__project_id__recipients__recipient_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "recipient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор получателя",
              "title": "Recipient Id"
            },
            "description": "Идентификатор получателя"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareRecipientUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareRecipientOut"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Данные запроса не прошли проверку схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/share-links": {
      "get": {
        "tags": [
          "выдача"
        ],
        "summary": "Выданные ссылки",
        "operationId": "list_share_links_api_projects__project_id__share_links_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Сколько записей вернуть",
              "default": 20,
              "title": "Limit"
            },
            "description": "Сколько записей вернуть"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Сколько записей пропустить",
              "default": 0,
              "title": "Offset"
            },
            "description": "Сколько записей пропустить"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareLinkListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "выдача"
        ],
        "summary": "Выдать ссылку получателю",
        "description": "Ссылка ведет на наш адрес и содержит отзываемый токен. Подписанная ссылка на объект хранилища наружу не выдается: ее нельзя отозвать.",
        "operationId": "create_share_link_api_projects__project_id__share_links_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareLinkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareLinkOut"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Состояние объекта изменилось: обновите данные и повторите.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Данные запроса не прошли проверку схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/share-links/{link_id}": {
      "delete": {
        "tags": [
          "выдача"
        ],
        "summary": "Отозвать ссылку",
        "description": "Идемпотентно: повторный отзыв уже отозванной ссылки завершается успехом.",
        "operationId": "revoke_share_link_api_projects__project_id__share_links__link_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор ссылки",
              "title": "Link Id"
            },
            "description": "Идентификатор ссылки"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{project_id}/export-bundles/{bundle_id}": {
      "get": {
        "tags": [
          "выдача"
        ],
        "summary": "Состояние архива",
        "operationId": "read_export_bundle_api_projects__project_id__export_bundles__bundle_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          },
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор архива",
              "title": "Bundle Id"
            },
            "description": "Идентификатор архива"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportBundleOut"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Конверт ошибки: любой ответ с кодом 4xx/5xx выглядит так."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShareLinkListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ShareLinkOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ShareLinkListResponse"
      },
...
      "ShareRecipientCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "role": {
            "$ref": "#/components/schemas/ShareRecipientRole"
          },
          "material": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Material"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "role"
        ],
        "title": "ShareRecipientCreate"
      },
      "ShareRecipientListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ShareRecipientOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ShareRecipientListResponse"
      },
      "ShareRecipientOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "$ref": "#/components/schemas/ShareRecipientRole"
          },
          "material": {
            "type": "string",
            "title": "Material",
            "description": "Что именно ему выдается, словами для экрана"
          },
          "status": {
            "$ref": "#/components/schemas/ShareRecipientStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "material",
          "status"
        ],
        "title": "ShareRecipientOut",
        "description": "Получатель материалов."
      },
...
      "ShareRecipientUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShareRecipientRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "material": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Material"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ShareRecipientUpdate"
      },
...
    }
  }
}